* Changed PrintInt() from s64 to s32, since Myriad compiler is missing some s64 functions * Put some benchmarking arrays in DDR * Updated GetTime() for Myriad (not tested) * Updated ldscript * Alphabetized embeddedCommon.h * Re-added benchmark unit test

+++ include/anki/embeddedCommon/utilities_c.h
@@ -67,7 +67,7 @@ extern "C" {
   void explicitPrintf(int reverseWords, const char *format, ...);
   // void explicitPrintfWithExplicitBuffer(int reverseWords, int * buffer, const char *format, ...);
 
-  void PrintInt(s64 value); // Print a single number
+  void PrintInt(s32 value); // Print a single number
   void PrintFloat(f64 value); // Print a single float
 
 #if defined(USING_MOVIDIUS_GCC_COMPILER)
